MDAFN_git/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.20.0
opencv-python>=4.5.0
pycocotools>=2.0.0
tqdm>=4.60.0
pyyaml>=5.4.0
matplotlib>=3.4.0
albumentations>=1.0.0
tensorboard>=2.5.0
pillow>=8.2.0
scipy>=1.7.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
timm>=0.4.12 